+++ index.ts
@@ -6,7 +6,7 @@ import { rmdir, stat, Stats, unlink } from 'fs';
 import { bindNodeCallback, empty, from, fromEvent, Observable, of, throwError } from 'rxjs';
 import { buffer, catchError, filter, map, mergeMap, switchMap, takeUntil, tap, toArray } from 'rxjs/operators';
 
-import program = require('commander');
+import program from 'commander';
 
 let untildify:(( path:string ) => string) = require('untildify');
 
@@ -38,7 +38,7 @@ function mdfind( name:string, excludeDirs:Array<RegExp> ):Observable<FileInfo> {
             .pipe( tap( err => console.error( "error", err ) ) )
             .pipe( mergeMap( err => throwError(err )));
         let onClose = fromEvent( mdfind, "close")
-            .pipe( tap( (value:any) => console.log( "closed", value[0] ) ));
+            //.pipe( tap( (value:any) => console.log( "closed", value[0] ) ));
 
         return fromEvent( mdfind.stdout, "data")
             .pipe(  takeUntil( onError ), 
@@ -66,7 +66,7 @@ function print( value:FileInfo ) {
         
 }
 
-function choice( fileInfo:FileInfo[] ):Observable<any>{
+function choice( fileInfo:FileInfo[], pageSize:number ):Observable<any>{
 
     let module = inquirer.createPromptModule( );
 
@@ -74,28 +74,37 @@ function choice( fileInfo:FileInfo[] ):Observable<any>{
             .filter( f => f.stats && (f.stats.isFile() || f.stats.isDirectory()) )
             .map( f => { return { name: f.path, value: f } } );
 
-    return from( module( { 
+    return from( module( [{ 
         name:"elements",
         type: "checkbox",
-        choices: choices
-    })) 
+        choices: choices,
+        pageSize: pageSize
+    }])) 
 }
 
 
 
-function remove( value:FileInfo ):Observable<FileInfo> {
+function remove( value:FileInfo, dryRun = false ):Observable<FileInfo> {
     if( value.stats ) {
-        if( value.stats.isFile() )
+        if( value.stats.isFile() ) {
+            console.log( `rm '${value.path}'`);
+            if( dryRun ) return empty();
             return rx_unlink( value.path ).pipe( map( v => value ));
-        else if( value.stats.isDirectory() )
-            return rx_exec( `rm -r  '${value.path}'` ).pipe( map( v => value ));  
+        }
+        else if( value.stats.isDirectory() ) {
+            console.log( `rm -r  '${value.path}'`);
+            if( dryRun ) return empty();
+            return rx_exec( `rm -r  '${value.path}'` ).pipe( map( v => value ));      
+        }
 
     } 
     return empty();
 
 }
 function clean( appName:string, option:any ) {
-
+    
+    process.stdout.write('\x1Bc');
+    
     console.log( appName );
 
     //console.log( 'clean ', appName, path.resolve( String(option.excludeDir) ) );
@@ -109,14 +118,24 @@ function clean( appName:string, option:any ) {
                             .map( p => { return new RegExp( '^' + p ) });
     }
 
+    const msg = () => {
+        if( option.dryRun ) 
+            console.log( "#\n# These files should be removed\n#");
+        else 
+            console.log( "#\n# These files have been removed\n#");
+    }
+
     mdfind( appName, excludeDirs)
-    //.pipe( tap( print ) )
-    .pipe( toArray(), map( files => files.sort( sortFileInfo ) ), switchMap( choice ) )
+    .pipe( toArray(), 
+        map( files => files.sort( sortFileInfo ) ), 
+        switchMap( files => choice(files, Number(option.pageSize)) ) )
+    .pipe( tap( msg ) )
     .pipe( mergeMap( v => from(v.elements) ))
-    .pipe( mergeMap( remove ) )
+    .pipe( mergeMap( f => remove(f, option.dryRun ) ) )
     .subscribe( 
-        v => console.log( "'%s' removed!", v.path),
-        err => console.error( err)
+        v => {},
+        err => console.error( err),
+        () => console.log('\n')
     );
 
 }
@@ -124,6 +143,8 @@ function clean( appName:string, option:any ) {
 program
     .version(_package.version, '-v --version')
     .option( "--excludeDir <dir[,dir,...]>", "exclude folder list")
+    .option( "--dryRun", "simulate execution (file will non be deleted)")
+    .option( "--pageSize <n>", "number of lines that will be shown per page", 10)
     .arguments( '<app name>' )
     .action( clean )
     .parse( process.argv);
